Move extensions and utilties into lower layers
