+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Core/Utilities/IntegerUtilities.cs
